+++ frontend/src/components/ui/StatisticsGraph/StatisticsGraph.stories.tsx
@@ -1,7 +1,7 @@
 import { getCongestionStatistics } from '@mocks/data';
 import type { Meta } from '@storybook/react';
 
-import { ENGLISH_DAYS } from '@constants';
+import { ENGLISH_DAYS } from '@constants/congestion';
 
 import StatisticsGraph from '.';
